+++ app/src/main/java/com/example/newfilmlistapp/ui/movie_detail/MovieDetailFragment.kt
@@ -17,7 +17,7 @@ import com.example.newfilmlistapp.view_model.MovieDetailViewModel
 import com.example.newfilmlistapp.databinding.FragmentMovieDetailBinding
 
 
-class MovieDetail : Fragment() {
+class MovieDetailFragment : Fragment() {
 
     private lateinit var binding: FragmentMovieDetailBinding
