+++ app/src/main/res/layout/fragment_movie_detail.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.movie_detail.MovieDetail">
+    tools:context=".ui.movie_detail.MovieDetailFragment">
 
     <include
         android:id="@+id/topNavBar"
