+++ app/src/main/res/layout/fragment_popular.xml
@@ -7,7 +7,7 @@
     android:id="@+id/main_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.popular.Popular"
+    tools:context=".ui.popular.PopularFragment"
     android:background="@color/black">
