+++ app/src/main/res/layout/test_moviedeatil.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.movie_detail.MovieDetail">
+    tools:context=".ui.movie_detail.MovieDetailFragment">
 
 <!--    <include-->
 <!--        android:id="@+id/topNavBar"-->
