+++ app/src/main/res/navigation/favorite_graph.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/favorite"
-        android:name="com.example.newfilmlistapp.ui.favorites.Favorites"
+        android:name="com.example.newfilmlistapp.ui.favorites.FavoritesFragment"
         android:label="@string/favorites_fragment" >
         <action
             android:id="@+id/action_favorite_to_movieDetail_favorite"
@@ -14,6 +14,6 @@
     </fragment>
     <fragment
         android:id="@+id/movieDetail"
-        android:name="com.example.newfilmlistapp.ui.movie_detail.MovieDetail"
+        android:name="com.example.newfilmlistapp.ui.movie_detail.MovieDetailFragment"
         android:label="MovieDetail" />
 </navigation>
