+++ app/src/main/res/navigation/popular_graph.xml
@@ -7,14 +7,14 @@
     <fragment
         android:id="@+id/popular"
         android:label="@string/popular_fragment"
-        android:name="com.example.newfilmlistapp.ui.popular.Popular" >
+        android:name="com.example.newfilmlistapp.ui.popular.PopularFragment" >
         <action
             android:id="@+id/action_popular_to_movieDetail_popular"
             app:destination="@id/movieDetail" />
     </fragment>
     <fragment
         android:id="@+id/movieDetail"
-        android:name="com.example.newfilmlistapp.ui.movie_detail.MovieDetail"
+        android:name="com.example.newfilmlistapp.ui.movie_detail.MovieDetailFragment"
         android:label="MovieDetail" />
 
 </navigation>
